pyproject.toml
[project]
name = "dio-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Luiz Eduardo",email = "luiz.rezendemoraes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "pyjwt"
]

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "*"
pytest = "*"
httpx = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
